Iteration 3 Delete useless code
